move the helper functions where they belong
